updated commit 1

+++ src/main/java/com/konic/library/Service/BookService.java
@@ -1,6 +1,7 @@
 package com.konic.library.Service;
 
 import com.konic.library.Entity.BookEntity;
+import com.konic.library.Exception.BookNotFoundException;
 import com.konic.library.Repository.BookRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,10 +32,13 @@ public class BookService {
     }
 
 
-public List<BookEntity> getAllBooks(){
+public List<BookEntity> getAllBooks() throws BookNotFoundException {
     log.info("Service: Fetching all books");
-    return bookrepository.findAll();
-
+    List<BookEntity> books = bookrepository.findAll();
+    if (books == null || books.isEmpty()) {
+        throw new BookNotFoundException("No books found in the library");
+    }
+    return books;
 }
 
 public BookEntity updateBook(BookEntity book){
@@ -54,16 +58,45 @@ public BookEntity updateBook(BookEntity book){
     return bookrepository.save(book);
 }
 
-public void deleteBook(Long id){
+public void deleteBook(Long id) throws BookNotFoundException {
+    log.info("Service: Deleting book with id: {}", id);
+
+    if (!bookrepository.existsById(id)) {
+        throw new BookNotFoundException("Book with ID " + id + " not found");
+    }
     log.info("Service: Deleting book with id: {}", id);
     bookrepository.deleteById(id);
 }
 
 public List<BookEntity> addMultipleBooks(List<BookEntity> books){
+    if (books == null || books.isEmpty()) {
+        throw new IllegalArgumentException("Book list cannot be empty");
+    }
+    for (BookEntity book : books) {
+        if (book.getTitle() == null || book.getTitle().trim().isEmpty()) {
+            throw new IllegalArgumentException("Book title cannot be empty");
+        }
+        if (book.getAuthor() == null || book.getAuthor().trim().isEmpty()) {
+            throw new IllegalArgumentException("Book author cannot be empty or Null");
+        }
+        if (book.getTotalcopies() <= 0) {
+            throw new IllegalArgumentException("Book totalcopies must be greater than zero");
+        }
+        book.setAvailablecopies(book.getTotalcopies());
+    }
     log.info("Service: Adding multiple books: {}", books);
     return bookrepository.saveAll(books);
 }
-public void deleteMultipleBooks(List<Long> ids){
+public void deleteMultipleBooks(List<Long> ids) throws BookNotFoundException {
+    if (ids == null || ids.isEmpty()) {
+        throw new IllegalArgumentException("Book ID list cannot be empty");
+    }
+    for (Long id : ids) {
+        if (!bookrepository.existsById(id)) {
+            throw new BookNotFoundException("Book with ID " + id + " not found");
+        }
+    }
+    log.info("Service: Deleting multiple books with IDs: {}", ids);
     bookrepository.deleteAllById(ids);
 }
